Made Amnon's suggested changes in hispull request review.
